requirements.txt
click==8.0.3
dataclasses==0.8
Flask==2.0.2
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
xlrd==2.0.1
et-xmlfile==1.1.0
openpyxl==3.0.9
mysqlclient==2.1.0
python-dotenv==0.19.2